merge

+++ src/main/java/com/mongle/api/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.mongle.api.repository;
 import com.mongle.api.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
 }
